Connect to ::apply-tag and ::remove-tag for the no-spell-check region
When the syntax highlighter update the no-spell-check region, we need to
listen to those changes.

The implementation can be improved to not get the tag name at each
signal emission, this will be done in a later commit for clarity.

The signal handlers are disconnected thanks to
g_signal_handlers_disconnect_matched() (already written for the other
signal handlers).